move utils functions to utils.js
